Закидываю контрольную весию ща буду эксперементировать.
