+++ lib/app/screens/Description/api/app_document_dir.dart
@@ -1,0 +1,4 @@
+import 'dart:io';
+import 'package:path_provider/path_provider.dart';
+
+Future<Directory> appDocumentDir() async => await getApplicationDocumentsDirectory();
