+++ lib/app/screens/Description/index.dart
@@ -1,3 +1,7 @@
 export './ui/description.dart';
 export './ui/popup_widget.dart';
+export './ui/content_block_widget.dart';
+export './ui/img_block_widget.dart';
 export './api/download_file.dart';
+export './api/app_document_dir.dart';
+export './api/hive_box_visible.dart';
