+++ lib/app/screens/Description/ui/description.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
-import 'package:hive/hive.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:flutter/material.dart';
+import 'package:hive/hive.dart';
 import '/app/router/export.dart';
 
 class Description extends StatefulWidget {
@@ -15,16 +14,10 @@ class Description extends StatefulWidget {
 class _DescriptionState extends State<Description> {
   Session get session => widget.session;
   bool fileExists = false;
-  bool isNotVisible = false;
 
-  Future<Directory> appDocumentDir() async => await getApplicationDocumentsDirectory();
-
-  Future<Box<dynamic>> getBox() async {
-    Directory appDocDir = await appDocumentDir();
-    Hive.init(appDocDir.path);
-    Box<dynamic> box = await Hive.openBox('visible');
-    return box;
-  }
+  late bool isNotVisible;
+  late String fileName;
+  late Box<dynamic> box;
 
   void toggleFileExist() {
     setState(() {
@@ -32,46 +25,16 @@ class _DescriptionState extends State<Description> {
     });
   }
 
-  bool isFileExists(String filePath) {
-    File file = File(filePath);
-    return file.existsSync();
-  }
-
-  String getFileName(track) {
-    RegExp regex = RegExp(r'\/([^\/]+)\.(mp3|mp4)$');
-    Match? match = regex.firstMatch(track);
-
-    if (match != null) {
-      String fileName = match.group(1)!;
-      return fileName;
-    }
-
-    return 'file_name';
-  }
-
-  String getFileExtantion(track) {
-    RegExp regex = RegExp(r'\.([a-zA-Z0-9]+)$');
-    Match? match = regex.firstMatch(track);
-
-    if (match != null) {
-      String extantion = match.group(1)!;
-      return extantion;
-    }
-
-    return 'mp3';
-  }
-
-  String get fileName => '${getFileName(session.track)}.${getFileExtantion(session.track)}';
-
   void init() async {
+    fileName = DownloadFile().getFileExt(widget.session.track); // name.ext
     Directory appDocDir = await appDocumentDir();
     String filePath = '${appDocDir.path}/$fileName';
-    fileExists = isFileExists(filePath);
+    fileExists = DownloadFile().isFileExists(filePath);
     setState(() {});
   }
 
   void initHive() async {
-    var box = await getBox();
+    box = await HiveBoxVisible().getBox();
 
     setState(() {
       isNotVisible = box.get('isNotVisible', defaultValue: false);
@@ -82,7 +45,7 @@ class _DescriptionState extends State<Description> {
   }
 
   void toggleIsNotVisible(boolean) async {
-    var box = await getBox();
+    box = await HiveBoxVisible().getBox();
 
     print('Проверка связи!');
     isNotVisible = boolean;
@@ -102,8 +65,6 @@ class _DescriptionState extends State<Description> {
 
   @override
   Widget build(BuildContext context) {
-    print('!fileExists && !isNotVisible 1 ${!fileExists && !isNotVisible}');
-
     return Scaffold(
       body: SafeArea(
         bottom: false,
@@ -133,9 +94,9 @@ class _DescriptionState extends State<Description> {
                   children: [
                     Column(
                       children: [
-                        _imgBlockWidget(session: session),
+                        imgBlockWidget(session: session),
                         const SizedBox(height: 15),
-                        _contentBlockWidget(session: session),
+                        contentBlockWidget(session: session),
                         const SizedBox(height: 10),
                         ElevatedButton(
                           onPressed: () {
@@ -200,58 +161,36 @@ class _DescriptionState extends State<Description> {
                         ),
                       ],
                     ),
-                    if (!fileExists && !isNotVisible)
-                      Positioned(
-                        right: 10,
-                        child: Container(
-                          decoration: const BoxDecoration(
-                            color: Color.fromARGB(255, 2, 52, 99),
-                            shape: BoxShape.circle,
-                          ),
-                          child: IconButton(
-                            padding: const EdgeInsets.all(0.0),
-                            iconSize: 40,
-                            icon: const Icon(
-                              Icons.download_for_offline_outlined,
-                            ),
-                            onPressed: () {
-                              showDialog(
-                                barrierDismissible: false,
-                                context: context,
-                                builder: (BuildContext context) {
-                                  return PopupDialog(
-                                      fileName: fileName,
-                                      session: session,
-                                      toggleParentFileExist: toggleFileExist,
-                                      toggleIsNotVisible: toggleIsNotVisible);
-                                },
-                              );
-                            },
-                            color: AppColors.mainColor,
-                          ),
+                    Positioned(
+                      right: 10,
+                      child: Container(
+                        decoration: const BoxDecoration(
+                          color: Color.fromARGB(255, 2, 52, 99),
+                          shape: BoxShape.circle,
                         ),
-                      )
-                    else if (!fileExists)
-                      Positioned(
-                        right: 10,
-                        child: Container(
-                          decoration: const BoxDecoration(
-                            color: Color.fromARGB(255, 2, 52, 99),
-                            shape: BoxShape.circle,
+                        child: IconButton(
+                          padding: const EdgeInsets.all(0.0),
+                          iconSize: 40,
+                          icon: const Icon(
+                            Icons.download_for_offline_outlined,
                           ),
-                          child: IconButton(
-                            padding: const EdgeInsets.all(0.0),
-                            iconSize: 40,
-                            icon: const Icon(
-                              Icons.download_for_offline_outlined,
-                            ),
-                            onPressed: () {
-                              print('object');
-                            },
-                            color: AppColors.mainColor,
-                          ),
+                          onPressed: () {
+                            showDialog(
+                              barrierDismissible: false,
+                              context: context,
+                              builder: (BuildContext context) {
+                                return PopupDialog(
+                                  fileName: fileName,
+                                  session: session,
+                                  toggleParentFileExist: toggleFileExist,
+                                );
+                              },
+                            );
+                          },
+                          color: AppColors.mainColor,
                         ),
-                      )
+                      ),
+                    )
                   ],
                 ),
               ),
@@ -269,87 +208,5 @@ class _DescriptionState extends State<Description> {
     } else {
       return VideoScreen(session: session);
     }
-  }
-}
-
-class _imgBlockWidget extends StatelessWidget {
-  const _imgBlockWidget({
-    required this.session,
-  });
-
-  final Session session;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width * 0.5,
-      padding: const EdgeInsets.symmetric(vertical: 8),
-      decoration: BoxDecoration(
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.47),
-            blurRadius: 4,
-            offset: const Offset(2, 3),
-          ),
-        ],
-        borderRadius: BorderRadius.circular(15.0),
-        color: AppColors.darkColor,
-      ),
-      child: Center(
-        child: Column(
-          children: [
-            FractionallySizedBox(
-              widthFactor: 0.7,
-              child: AspectRatio(
-                aspectRatio: 1,
-                child: Container(
-                  decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    image: DecorationImage(
-                      image: NetworkImage(session.sessionImg),
-                      fit: BoxFit.cover,
-                    ),
-                  ),
-                ),
-              ),
-            ),
-            Text(
-              session.subscribe.trim().replaceAll("\\n", "\n"),
-              textAlign: TextAlign.center,
-              style: const TextStyle(
-                fontFamily: FontFamily.regularFont,
-                fontSize: 16,
-                height: 1.3,
-              ),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _contentBlockWidget extends StatelessWidget {
-  const _contentBlockWidget({
-    required this.session,
-  });
-
-  final Session session;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width * 1.0,
-      color: Colors.white.withOpacity(0.8),
-      padding: const EdgeInsets.all(8.0),
-      child: Text(
-        session.description.trim().replaceAll("\\n", "\n"),
-        textAlign: TextAlign.left,
-        style: const TextStyle(
-          fontSize: 20.0,
-          fontFamily: FontFamily.regularFont,
-        ),
-      ),
-    );
   }
 }
